moved CollectCLIArgs to config package
