+++ svc-aggregation/main.go
@@ -49,6 +49,8 @@ func main() {
 		log.Fatal("error while trying to set configuration: " + err.Error())
 	}
 
+	config.CollectCLArgs()
+
 	if err := dc.SetConfiguration(config.Data.MessageQueueConfigFilePath); err != nil {
 		log.Fatal("error while trying to set messagebus configuration: " + err.Error())
 	}
