+++ svc-api/main.go
@@ -140,6 +140,8 @@ func main() {
 		log.Fatal(err.Error())
 	}
 
+	config.CollectCLArgs()
+
 	err = services.InitializeService(services.GoMicro, services.APIClient)
 	if err != nil {
 		log.Fatal("service initialisation failed: " + err.Error())
